add show time function in all page of job
